Move common gcs utils out of gcs reporter.
