+++ prow/crier/reporters/gcs/internal/util/BUILD.bazel
@@ -2,17 +2,17 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["reporter.go"],
-    importpath = "k8s.io/test-infra/prow/crier/reporters/gcs",
+    srcs = ["util.go"],
+    importpath = "k8s.io/test-infra/prow/crier/reporters/gcs/internal/util",
     visibility = ["//visibility:public"],
     deps = [
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
-        "//prow/crier/reporters/gcs/internal/util:go_default_library",
-        "//prow/errorutil:go_default_library",
-        "@com_github_googlecloudplatform_testgrid//metadata:go_default_library",
+        "//prow/gcsupload:go_default_library",
+        "//prow/pod-utils/downwardapi:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go//storage:go_default_library",
+        "@org_golang_google_api//googleapi:go_default_library",
     ],
 )
 
@@ -25,23 +25,19 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//prow/crier/reporters/gcs/internal/util:all-srcs",
-    ],
+    srcs = [":package-srcs"],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["reporter_test.go"],
+    srcs = ["util_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
-        "@com_github_google_go_cmp//cmp:go_default_library",
-        "@com_github_googlecloudplatform_testgrid//metadata:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
+        "@org_golang_google_api//googleapi:go_default_library",
     ],
 )
